requirements.txt
requests>=2.31.0
pandas>=2.0.0
pyarrow>=14.0.0
pyyaml>=6.0.0
pytest>=7.0.0
